pyproject.toml
[project]
name = "cover-maker"
version = "0.1.0"
description = "DevBcn Speaker Card Generator"
authors = [
    { name = "DevBcn Team" }
]
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "rembg[cpu]>=2.0",
    "Pillow>=10.0",
    "python-dotenv>=1.0",
    "google-api-python-client>=2.100",
    "google-auth>=2.23",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "responses>=0.25",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ANN"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN"]

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=90"
testpaths = ["tests"]
